[NCFO] Add support for downloading performance recordings.
Only RD2014 is supported for now...

+++ net-slurp/ncfo/rebuild-symlinks.sh
@@ -1,8 +1,14 @@
 #!/bin/sh
 
+TOP=/home/bert/scripts/net-slurp/ncfo
+
 find_top_level_symlinks () {
     if [ "$#" -eq 0 ]; then set -- -print; fi
     find . -name . -o -type d -prune -o -type l "$@"
+}
+
+get_subdirs () {
+    echo `ls -l "$TOP" | sed -e '/^d/!d;s/.* //;s/$/ |/'` | sed -e 's/ |$//'
 }
 
 if [ $# -lt 1 ]; then
@@ -19,20 +25,20 @@ if [ $# -lt 1 ]; then
 fi
 
 case "$1" in
-    opera)
-	subdir=opera ;;
-    science|chorus)
-	subdir=science ;;
-    *)
-	echo "If guessing is not possible, you must specify an argument!" >&2
-	echo "" >&2
-	echo "Usage: `basename "$0"` [opera | science]" >&2
-	exit 1 ;;
+    chorus) subdir=science ;;
+    *)      subdir="$1" ;;
 esac
+
+if [ -z "$subdir" -o ! -d "$TOP/$subdir" ]; then
+    echo "If guessing is not possible, you must specify an argument!" >&2
+    echo "" >&2
+    echo "Usage: `basename "$0"` [`get_subdirs`]" >&2
+    exit 1
+fi
 
 # Remove symlinks in the current directory, but not in subdirectories.
 find_top_level_symlinks -print0 | xargs -0 -r rm -f
 
 # Recreate symlinks.
-ln -s /home/bert/scripts/net-slurp/ncfo{,/$subdir}/*.*[^~] .
+ln -s "$TOP"{,/"$subdir"}/*.*[^~] .
 ln -s /home/bert/scripts/net-slurp/plinks.pl .
